Revert "[web] consolidate JS interop code (#30007)"
This reverts commit a104af8a2916c1611e889b69558388267eac26ff.
