+++ lib/web_ui/lib/src/engine/canvaskit/canvaskit_api.dart
@@ -2162,64 +2162,3 @@ class SkPartialImageInfo {
   external int get height;
   external int get width;
 }
-
-/// Monkey-patch the top-level `module` and `exports` objects so that
-/// CanvasKit doesn't attempt to register itself as an anonymous module.
-///
-/// The idea behind making these fake `exports` and `module` objects is
-/// that `canvaskit.js` contains the following lines of code:
-///
-///     if (typeof exports === 'object' && typeof module === 'object')
-///       module.exports = CanvasKitInit;
-///     else if (typeof define === 'function' && define['amd'])
-///       define([], function() { return CanvasKitInit; });
-///
-/// We need to avoid hitting the case where CanvasKit defines an anonymous
-/// module, since this breaks RequireJS, which DDC and some plugins use.
-/// Temporarily removing the `define` function won't work because RequireJS
-/// could load in between this code running and the CanvasKit code running.
-/// Also, we cannot monkey-patch the `define` function because it is
-/// non-configurable (it is a top-level 'var').
-// TODO(hterkelsen): Rather than this monkey-patch hack, we should
-// build CanvasKit ourselves. See:
-// https://github.com/flutter/flutter/issues/52588
-void patchCanvasKitModule(html.ScriptElement canvasKitScript) {
-  // First check if `exports` and `module` are already defined. If so, then
-  // CommonJS is being used, and we shouldn't have any problems.
-  final js.JsFunction objectConstructor = js.context['Object'] as js.JsFunction;
-  if (js.context['exports'] == null) {
-    final js.JsObject exportsAccessor = js.JsObject.jsify(<String, dynamic>{
-      'get': allowInterop(() {
-        if (html.document.currentScript == canvasKitScript) {
-          return js.JsObject(objectConstructor);
-        } else {
-          return js.context['_flutterWebCachedExports'];
-        }
-      }),
-      'set': allowInterop((dynamic value) {
-        js.context['_flutterWebCachedExports'] = value;
-      }),
-      'configurable': true,
-    });
-    objectConstructor.callMethod(
-        'defineProperty', <dynamic>[js.context, 'exports', exportsAccessor]);
-  }
-  if (js.context['module'] == null) {
-    final js.JsObject moduleAccessor = js.JsObject.jsify(<String, dynamic>{
-      'get': allowInterop(() {
-        if (html.document.currentScript == canvasKitScript) {
-          return js.JsObject(objectConstructor);
-        } else {
-          return js.context['_flutterWebCachedModule'];
-        }
-      }),
-      'set': allowInterop((dynamic value) {
-        js.context['_flutterWebCachedModule'] = value;
-      }),
-      'configurable': true,
-    });
-    objectConstructor.callMethod(
-        'defineProperty', <dynamic>[js.context, 'module', moduleAccessor]);
-  }
-  html.document.head!.append(canvasKitScript);
-}
