+++ lib/web_ui/lib/src/engine/profiler.dart
@@ -4,11 +4,14 @@
 
 import 'dart:async';
 import 'dart:html' as html;
+import 'dart:js_util' as js_util;
 
 import 'package:ui/ui.dart' as ui;
 
 import 'platform_dispatcher.dart';
-import 'safe_browser_api.dart';
+
+/// A function that receives a benchmark [value] labeleb by [name].
+typedef OnBenchmark = void Function(String name, double value);
 
 /// A function that computes a value of type [R].
 ///
@@ -102,16 +105,12 @@ class Profiler {
   void benchmark(String name, double value) {
     _checkBenchmarkMode();
 
-    // First get the value as `Object?` then use `as` cast to check the type.
-    // This is because the type cast in `getJsProperty<Object?>` is optimized
-    // out at certain optimization levels in dart2js, leading to obscure errors
-    // later on.
-    final Object? onBenchmark = getJsProperty<Object?>(
-      html.window,
-      '_flutter_internal_on_benchmark',
-    );
-    onBenchmark as OnBenchmark?;
-    onBenchmark?.call(name, value);
+    final OnBenchmark? onBenchmark =
+        // ignore: implicit_dynamic_function
+        js_util.getProperty(html.window, '_flutter_internal_on_benchmark') as OnBenchmark?;
+    if (onBenchmark != null) {
+      onBenchmark(name, value);
+    }
   }
 }
