+++ lib/web_ui/lib/src/engine/text/font_collection.dart
@@ -5,11 +5,11 @@
 import 'dart:async';
 import 'dart:convert';
 import 'dart:html' as html;
+import 'dart:js_util' as js_util;
 import 'dart:typed_data';
 
 import '../assets.dart';
 import '../browser_detection.dart';
-import '../safe_browser_api.dart';
 import '../util.dart';
 import 'layout_service.dart';
 
@@ -185,7 +185,18 @@ class FontManager {
     try {
       final html.FontFace fontFace = html.FontFace(family, asset, descriptors);
       _fontLoadingFutures.add(fontFace.load().then((_) {
-        html.document.fonts!.add(fontFace);
+        // We could do:
+        // ```
+        // html.document.fonts!.add(fontFace);
+        // ```
+        // But dart:html expects the return value to be non-null, and Firefox
+        // returns null. This causes the app to crash in Firefox with a null
+        // check exception.
+        //
+        // TODO(mdebbar): Revert this once the dart:html type is fixed.
+        //                https://github.com/dart-lang/sdk/issues/45676
+        // ignore: implicit_dynamic_function
+        js_util.callMethod(html.document.fonts!, 'add', <dynamic>[fontFace]);
       }, onError: (dynamic e) {
         printWarning('Error while trying to load font family "$family":\n$e');
       }));
@@ -312,3 +323,9 @@ class _PolyfillFontManager extends FontManager {
     _fontLoadingFutures.add(completer.future);
   }
 }
+
+final bool supportsFontLoadingApi =
+    js_util.hasProperty(html.window, 'FontFace');
+final bool supportsFontsClearApi =
+    js_util.hasProperty(html.document, 'fonts') &&
+        js_util.hasProperty(html.document.fonts!, 'clear');
