+++ lib/web_ui/test/engine/surface/scene_builder_test.dart
@@ -921,6 +921,7 @@ HtmlImage createTestImage({int width = 100, int height = 50}) {
   ctx.fillRect(66, 0, 33, 50);
   ctx.fill();
   final html.ImageElement imageElement = html.ImageElement();
-  imageElement.src = js_util.callMethod<String>(canvas, 'toDataURL', <dynamic>[]);
+  // ignore: implicit_dynamic_function
+  imageElement.src = js_util.callMethod(canvas, 'toDataURL', <dynamic>[]) as String;
   return HtmlImage(imageElement, width, height);
 }
