+++ lib/web_ui/test/html/drawing/canvas_draw_image_golden_test.dart
@@ -734,7 +734,8 @@ HtmlImage createTestImage({int width = 100, int height = 50}) {
   ctx.fillRect(66, 0, 33, 50);
   ctx.fill();
   final html.ImageElement imageElement = html.ImageElement();
-  imageElement.src = js_util.callMethod<String>(canvas, 'toDataURL', <dynamic>[]);
+  // ignore: implicit_dynamic_function
+  imageElement.src = js_util.callMethod(canvas, 'toDataURL', <dynamic>[]) as String;
   return HtmlImage(imageElement, width, height);
 }
