+++ lib/web_ui/test/html/drawing/draw_vertices_golden_test.dart
@@ -424,6 +424,7 @@ Future<HtmlImage> createTestImage({int width = 50, int height = 40}) {
   imageElement.onLoad.listen((html.Event event) {
     completer.complete(HtmlImage(imageElement, width, height));
   });
-  imageElement.src = js_util.callMethod<String>(canvas, 'toDataURL', <dynamic>[]);
+  // ignore: implicit_dynamic_function
+  imageElement.src = js_util.callMethod(canvas, 'toDataURL', <dynamic>[]) as String;
   return completer.future;
 }
