+++ lib/web_ui/test/paragraph_builder_test.dart
@@ -4,6 +4,7 @@
 
 import 'package:test/bootstrap/browser.dart';
 import 'package:test/test.dart';
+import 'package:ui/src/engine.dart';
 import 'package:ui/ui.dart';
 
 void main() {
@@ -11,6 +12,10 @@ void main() {
 }
 
 void testMain() {
+  setUpAll(() {
+    WebExperiments.ensureInitialized();
+  });
+
   test('Should be able to build and layout a paragraph', () {
     final ParagraphBuilder builder = ParagraphBuilder(ParagraphStyle());
     builder.addText('Hello');
